Description: This is a Python program that allows the user to choose between two programs: ROT13 encryption/decryption and visual cryptography.
In the ROT13 program, the user can input a message and choose between encryption or decryption mode. If encryption mode is chosen, the user is prompted to input a shift value, and each alphabetic character in the message is shifted by the given shift value. If decryption mode is chosen, the user is prompted to input the shift value used to encrypt the message, and each alphabetic character in the message is shifted by the inverse of the given shift value. Non-alphabetic characters are left as-is.

In the visual cryptography program, the program loads two image files ("scrambled1.png" and "scrambled2.png"), scrambles their pixels together, and saves the resulting image as "output.png".

Dependencies:
This program requires the following modules to be installed:

pyttsx3 (for text-to-speech functionality)
PIL (for image manipulation)
Usage:
Run the Python program using a Python interpreter such as IDLE or the command line.
Choose between ROT13 encryption/decryption and visual cryptography by entering "R" or "V", respectively.
If ROT13 encryption/decryption is chosen, follow the prompts to input a message, choose a mode (encryption or decryption), and input a shift value (if encrypting).
If visual cryptography is chosen, ensure that the two image files "scrambled1.png" and "scrambled2.png" are located in the same directory as the Python program. If not, an error message will be displayed.
After the program has finished, the resulting encrypted/decrypted message or the output image will be displayed and/or spoken aloud (if text-to-speech is enabled).